+++ src/stepZ_continuation.test.ts
@@ -1,4 +1,4 @@
-import Interpreter, { isContinuation } from "./interpreter";
+import Interpreter, { isContinuation, ContinuablePromise } from "./interpreter";
 
 const DEBUG = false;
 const DEBUG_MAX = 100;
@@ -144,4 +144,132 @@ test("Check if JS method can not raise any continuation.", () => {
       [".", "obj", ["`", "method"]]
     ])
   ).toThrow(/Javascript methods can not throw any continuation/);
+});
+
+test("Check if await throws continuation.", () => {
+  expect(() =>
+    interpreter.eval(["await"])
+  ).toThrow(ContinuablePromise);
+});
+
+test("Check if await can resume the process.", () => {
+  try {
+    interpreter.eval(["do", ["await"], 10])
+    throw new Error('failed');
+  } catch (e) {
+    if (e instanceof ContinuablePromise) {
+      expect(e.resume()).toBe(10);
+    } else {
+      throw new Error('failed');
+    }
+  }
+});
+
+test("Check if await can receive and return the resolved value", async () => {
+  try {
+    interpreter.eval(["do", ["await", Promise.resolve(15)]])
+    throw new Error('failed');
+  } catch (e) {
+    if (e instanceof ContinuablePromise) {
+      await e;
+      expect(e.resume()).toBe(15);
+    } else {
+      throw new Error('failed');
+    }
+  }
+});
+
+test("Check if await can receive and return the resolved value", async () => {
+  try {
+    interpreter.eval(["do", ["await", 100]])
+    throw new Error('failed');
+  } catch (e) {
+    if (e instanceof ContinuablePromise) {
+      await e;
+      expect(e.resume()).toBe(100);
+    } else {
+      throw new Error('failed');
+    }
+  }
+});
+
+test("Check ContinuationPromise status methods work when pending.", async () => {
+  try {
+    interpreter.eval(["await", new Promise(resolve=>setTimeout(()=>null, 1000))]);
+    throw new Error('failed');
+  } catch (e) {
+    if (e instanceof ContinuablePromise) {
+      expect(e.isPending()).toBe(true);
+      expect(e.isFulfilled()).toBe(false);
+      expect(e.isRejected()).toBe(false);
+      expect(()=>e.reason()).toThrow();
+    } else {
+      throw new Error('failed');
+    }
+  }
+});
+
+test("Check ContinuationPromise status methods work when resolved.", async () => {
+  try {
+    interpreter.eval(["await", Promise.resolve(10)]);
+    throw new Error('failed');
+  } catch (e) {
+    if (e instanceof ContinuablePromise) {
+      await e;
+      expect(e.isPending()).toBe(false);
+      expect(e.isFulfilled()).toBe(true);
+      expect(e.isRejected()).toBe(false);
+      expect(()=>e.reason()).toThrow();
+    } else {
+      throw new Error('failed');
+    }
+  }
+});
+
+test("Check ContinuationPromise status methods work when rejected.", async () => {
+  try {
+    interpreter.eval(["await", Promise.reject(14)]);
+    throw new Error('failed');
+  } catch (e) {
+    if (e instanceof ContinuablePromise) {
+      await null;
+      expect(e.isPending()).toBe(false);
+      expect(e.isFulfilled()).toBe(false);
+      expect(e.isRejected()).toBe(true);
+      expect(e.reason()).toBe(14);
+      try {
+        await e;
+      } catch (e) {
+        // Nothing. Jest may throw an error some other test after this test
+        // if we don't await here...Maybe this is a Jest's bug? ⇒ Maybe so.
+        // Even if we do Promise.reject(xxx) outside interpreter.eval()
+        // in test(), this phenomenon occures.
+      }
+      // Promise.reject("XXXX"); // If you uncomment this, error occures some other test!
+    } else {
+      throw new Error('failed');
+    }
+  }
+});
+
+test("Check if evalAsync works.", async () => {
+  expect(await interpreter.evalAsync(["do", ["await", Promise.resolve()], 555])).toBe(555);
+});
+
+
+test("Check if evalAsync and await really waits promise.", async () => {
+  const start = performance.now();
+  await interpreter.evalAsync(["await", new Promise(resolve=>setTimeout(resolve, 1000))]);
+  const duration = performance.now() - start;
+  expect(duration).toBeGreaterThan(800);
+  expect(duration).toBeLessThan(1200);
+});
+
+test("Check if evalAsync and await can handle rejected promise.", async () => {
+  try {
+    await interpreter.evalAsync(["await", Promise.reject("error")]);
+    throw new Error("failed");
+  } catch (e) {
+    expect(e).toBe("error");
+  }
 });